requirements.txt
tinydb
sanic
sanic-jinja2
netaddr
pycryptodome
